build 1

+++ src/main/java/com/ecommerce/repository/CountryRepository.java
@@ -8,6 +8,6 @@ import com.ecommerce.entity.Country;
 
 @CrossOrigin("http://localhost:4200")
 @RepositoryRestResource(collectionResourceRel = "countries", path = "countries")
-public interface CountryRepository extends JpaRepository<Country, Integer> {
+public interface CountryRepository extends JpaRepository<Country, Long> {
     
 }
